refactor(runtimed): move env progress projection into package
